Painful memory consuming error solved, blessed by the glorious mlx_sync

+++ cub3d_utils_2.c
@@ -6,7 +6,7 @@
 /*   By: fgata-va <fgata-va@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/10 11:34:02 by fgata-va          #+#    #+#             */
-/*   Updated: 2021/01/31 23:11:17 by fgata-va         ###   ########.fr       */
+/*   Updated: 2021/02/02 12:36:56 by fgata-va         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,9 +52,10 @@ int		ft_isnumber(char *s, int id)
 int		end_program(t_args *game_data)
 {
 	ft_free_all(game_data->data, game_data->tex, game_data->file);
-	if (BONUS)
+	if (BONUS && game_data->data->music_process >= 0)
 		kill(game_data->data->music_process, 9);
 	ft_printf("See you, space cowboy\n");
+	system("leaks --fullContent cub3D");
 	exit(0);
 	return (0);
 }
